# swagger.json
{
	"components": {
...
		"schemas": {
			"Order": {
				"description": "Esquema con el cual el backend retorna un pedido.",
				"properties": {
					"_id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deliveryTime": {
						"type": "string",
						"format": "date-time"
					},
					"orderRating": {
						"type": "number",
						"format": "double"
					},
					"products": {
						"items": {
							"properties": {
								"quantity": {
									"type": "number",
									"format": "double"
								},
								"product": {
									"type": "string"
								}
							},
							"required": [
								"quantity",
								"product"
							],
							"type": "object"
						},
						"type": "array"
					},
					"disabled": {
						"type": "boolean"
					},
					"restaurant": {
						"type": "string"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"_v": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"createdAt",
					"updatedAt",
					"deliveryTime",
					"products",
					"disabled",
					"restaurant",
					"total",
					"user",
					"state",
					"_v"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"_id": "6531a1b0e9c33a3af568e47e",
					"createdAt": "2023-10-18T04:03:42.164Z",
					"updatedAt": "2023-10-20T04:03:42.164Z",
					"deliveryTime": "2030-10-20T04:03:42.164Z",
					"orderRating": 4.8,
					"products": [
						{
							"product": "6530777d23e50c837de795d4",
							"quantity": 3
						}
					],
					"restaurant": "65307280beb5043d524138d3",
					"user": "6531618e6025da22956875a6",
					"total": 100,
					"state": {
						"_id": "65305444746510934b074dff",
						"name": "Entregado"
					},
					"disabled": false,
					"_v": 0
				}
			},
...
			"OrderCreation": {
				"description": "Esquema para crear un pedido.",
				"properties": {
					"deliveryTime": {
						"type": "string",
						"format": "date-time"
					},
					"products": {
						"items": {
							"properties": {
								"quantity": {
									"type": "number",
									"format": "double"
								},
								"product": {
									"type": "string"
								}
							},
							"required": [
								"quantity",
								"product"
							],
							"type": "object"
						},
						"type": "array"
					},
					"restaurant": {
						"type": "string"
					}
				},
				"required": [
					"deliveryTime",
					"products",
					"restaurant"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrderUpdate": {
				"description": "Esquema para editar un pedido.",
				"properties": {
					"_id": {
						"type": "string"
					},
					"deliveryTime": {
						"type": "string",
						"format": "date-time"
					},
					"orderRating": {
						"type": "number",
						"format": "double"
					},
					"products": {
						"items": {
							"properties": {
								"quantity": {
									"type": "number",
									"format": "double"
								},
								"product": {
									"type": "string"
								}
							},
							"required": [
								"quantity",
								"product"
							],
							"type": "object"
						},
						"type": "array"
					},
					"user": {
						"type": "string"
					},
					"state": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"_id": "6531a1b0e9c33a3af568e47e",
					"deliveryTime": "2030-10-20T04:03:42.164Z",
					"orderRating": 4.8,
					"products": [
						{
							"product": "6530777d23e50c837de795d4",
							"quantity": 3
						}
					],
					"user": "6531618e6025da22956875a6",
					"state": "65305444746510934b074dff"
				}
			},
			"IMessage": {
				"description": "Respuesta simple con mensaje.",
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"status": 404,
					"message": "NOT FOUND"
				}
			},
			"Product": {
				"description": "Esquema con el cual el backend retorna un producto.",
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"category": {
						"properties": {
							"name": {
								"type": "string"
							},
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"name",
							"_id"
						],
						"type": "object"
					},
					"restaurant": {
						"type": "string"
					},
					"cost": {
						"type": "number",
						"format": "double"
					},
					"disabled": {
						"type": "boolean"
					},
					"_v": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"name",
					"description",
					"image",
					"category",
					"restaurant",
					"cost",
					"disabled",
					"_v"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"_id": "6530777d23e50c837de795d4",
					"name": "Giga Mac",
					"description": "45000 kCal y viaje garantizado al hospital",
					"image": "https://th.bing.com/th/id/OIP.JpDTcXVjo5p_CGPXJjcOwgHaE8?pid=ImgDet&rs=1",
					"category": {
						"_id": "65305444746510934b074df9",
						"name": "Hamburguesas"
					},
					"restaurant": "65307280beb5043d524138d3",
					"cost": 30,
					"disabled": false,
					"_v": 0
				}
			},
...
			"ProductCreation": {
				"description": "Esquema para crear un producto.",
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"restaurant": {
						"type": "string"
					},
					"cost": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"description",
					"image",
					"category",
					"restaurant",
					"cost"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"name": "Giga Mac",
					"description": "45000 kCal y viaje garantizado al hospital",
					"image": "https://th.bing.com/th/id/OIP.JpDTcXVjo5p_CGPXJjcOwgHaE8?pid=ImgDet&rs=1",
					"category": "65305444746510934b074df9",
					"restaurant": "65307280beb5043d524138d3",
					"cost": 30
				}
			},
			"ProductUpdate": {
				"description": "Esquema para editar un producto.",
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"restaurant": {
						"type": "string"
					},
					"cost": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"_id": "6530777d23e50c837de795d4",
					"name": "Giga Mac",
					"description": "45000 kCal y viaje garantizado al hospital",
					"image": "https://th.bing.com/th/id/OIP.JpDTcXVjo5p_CGPXJjcOwgHaE8?pid=ImgDet&rs=1",
					"category": "65305444746510934b074df9",
					"restaurant": "65307280beb5043d524138d3",
					"cost": 30
				}
			},
			"Restaurant": {
				"description": "Esquema con el cual el backend retorna un restaurante.",
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"administrator": {
						"type": "string"
					},
					"deliveryTime": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"properties": {
							"name": {
								"type": "string"
							},
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"name",
							"_id"
						],
						"type": "object"
					},
					"sales": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"disabled": {
						"type": "boolean"
					},
					"_v": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"name",
					"administrator",
					"deliveryTime",
					"category",
					"sales",
					"rating",
					"disabled",
					"_v"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"_id": "65307280beb5043d524138d3",
					"name": "McDonalds",
					"disabled": false,
					"administrator": "6531618e6025da22956875a6",
					"category": {
						"_id": "65305444746510934b074df9",
						"name": "Comida Rápida"
					},
					"deliveryTime": 30,
					"sales": 10,
					"rating": 4.7,
					"__v": 0
				}
			},
...
			"RestaurantCreation": {
				"description": "Esquema para crear un restaurante.",
				"properties": {
					"name": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"deliveryTime": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"category",
					"deliveryTime"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"name": "McDonalds",
					"category": "65305444746510934b074df9",
					"deliveryTime": 30
				}
			},
			"RestaurantUpdate": {
				"description": "Esquema para editar un restaurante.",
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"administrator": {
						"type": "string"
					},
					"deliveryTime": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"_id": "65307280beb5043d524138d3",
					"name": "McDonalds",
					"administrator": "6531618e6025da22956875a6",
					"category": "65305444746510934b074df9",
					"deliveryTime": 30
				}
			},
...
		},
		"securitySchemes": {
			"token": {
				"type": "http",
				"description": "JWT token based security",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"scopes": {
					"create": "Create a resource",
					"update": "Update a resource",
					"delete": "Delete a resource"
				}
			}
		}
	},
	"info": {
		"title": "Proyecto 1 Backend",
		"version": "1.0.0",
		"description": "API para ofrecer servicios de domicilios de restaurantes.",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "david"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/orders": {
			"get": {
				"operationId": "GetOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Order"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Obtiene todos los pedidos que cumplan con los filtros.",
				"summary": "Get Orders",
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"description": "Búsqueda por id del usuario que creó los pedidos",
						"in": "query",
						"name": "user",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Búsqueda por id de restaurante al cual se le hicieron los pedidos",
						"in": "query",
						"name": "restaurant",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Se retornarán los pedidos creados despues de esta fecha",
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"description": "Se retornarán los pedidos creados antes de esta fecha",
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Order"
								}
							}
						}
					}
				},
				"description": "Crea un nuevo pedido en la base de datos.",
				"summary": "Create Order",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"token": [
							"create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Datos del nuevo pedido a crear",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OrderCreation",
								"description": "Datos del nuevo pedido a crear"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Order"
								}
							}
						}
					}
				},
				"description": "Edita los datos de un pedido. El pedido es identificado por la propiedad _id.",
				"summary": "Update Order",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"token": [
							"update"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "El pedido a editar. Todos los campos que estén definidos en este objeto se sobreescriben.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OrderUpdate",
								"description": "El pedido a editar. Todos los campos que estén definidos en este objeto se sobreescriben."
							}
						}
					}
				}
			}
		},
		"/orders/unconfirmed": {
			"get": {
				"operationId": "GetUnconfirmed",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Order"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Obtiene todos los pedidos sin aceptar por el restaurante que cumplan con los filtros.",
				"summary": "Get Unconfirmed Orders",
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"description": "Búsqueda por id del usuario que creó los pedidos",
						"in": "query",
						"name": "user",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Búsqueda por id de restaurante al cual se le hicieron los pedidos",
						"in": "query",
						"name": "restaurant",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Se retornarán los pedidos creados despues de esta fecha",
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"description": "Se retornarán los pedidos creados antes de esta fecha",
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			}
		},
...
		"/orders/{id}": {
			"get": {
				"operationId": "GetOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Order"
								}
							}
						}
					}
				},
				"description": "Obtiene los datos de un pedido.",
				"summary": "Get Order",
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"description": "Id del pedido",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMessage"
								}
							}
						}
					}
				},
				"description": "Deshabilita un pedido, este ya no se podrá leer.",
				"summary": "Delete Order",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"token": [
							"delete"
						]
					}
				],
				"parameters": [
					{
						"description": "Id del pedido.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Product"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Obtiene todos los productos que cumplan con los filtros.",
				"summary": "Get Products",
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"description": "Búsqueda por id de restaurante",
						"in": "query",
						"name": "restaurant",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Búsqueda por id o nombre de categoría",
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				},
				"description": "Crea un nuevo producto en la base de datos.",
				"summary": "Create Product",
				"tags": [
					"Products"
				],
				"security": [
					{
						"token": [
							"create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Datos del nuevo producto a crear",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductCreation",
								"description": "Datos del nuevo producto a crear"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				},
				"description": "Edita los datos de un producto. El producto es identificado por la propiedad _id.",
				"summary": "Update Product",
				"tags": [
					"Products"
				],
				"security": [
					{
						"token": [
							"update"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "El producto a editar. Todos los campos que estén definidos en este objeto se sobreescriben.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductUpdate",
								"description": "El producto a editar. Todos los campos que estén definidos en este objeto se sobreescriben."
							}
						}
					}
				}
			}
		},
...
		"/products/{id}": {
			"get": {
				"operationId": "GetProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				},
				"description": "Obtiene los datos de un producto.",
				"summary": "Get Product",
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"description": "Id del producto",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMessage"
								}
							}
						}
					}
				},
				"description": "Deshabilita un producto, este ya no se podrá leer.",
				"summary": "Delete Product",
				"tags": [
					"Products"
				],
				"security": [
					{
						"token": [
							"delete"
						]
					}
				],
				"parameters": [
					{
						"description": "Id del producto.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/restaurants": {
			"get": {
				"operationId": "GetRestaurants",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Restaurant"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Obtiene todos los restaurantes que cumplan con los filtros.",
				"summary": "Get Restaurants",
				"tags": [
					"Restaurants"
				],
				"security": [],
				"parameters": [
					{
						"description": "Búsqueda por nombre de restaurantes",
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Búsqueda por id o nombre de categoría",
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Restaurant"
								}
							}
						}
					}
				},
				"description": "Crea un nuevo restaurante en la base de datos.",
				"summary": "Create Restaurant",
				"tags": [
					"Restaurants"
				],
				"security": [
					{
						"token": [
							"create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RestaurantCreation"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Restaurant"
								}
							}
						}
					}
				},
				"description": "Edita los datos de un restaurant. El restaurant es identificado por la propiedad _id.",
				"summary": "Update Restaurant",
				"tags": [
					"Restaurants"
				],
				"security": [
					{
						"token": [
							"update"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RestaurantUpdate"
							}
						}
					}
				}
			}
		},
...
		"/restaurants/{id}": {
			"get": {
				"operationId": "GetRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Restaurant"
								}
							}
						}
					}
				},
				"description": "Obtiene los datos de un restaurante.",
				"summary": "Get Restaurant",
				"tags": [
					"Restaurants"
				],
				"security": [],
				"parameters": [
					{
						"description": "Id del restaurante",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMessage"
								}
							}
						}
					}
				},
				"description": "Deshabilita un restaurante, este ya no se podrá leer.",
				"summary": "Delete Restaurant",
				"tags": [
					"Restaurants"
				],
				"security": [
					{
						"token": [
							"delete"
						]
					}
				],
				"parameters": [
					{
						"description": "Id del restaurante.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}